+++ corehq/apps/app_manager/util.py
@@ -1,4 +1,5 @@
 from collections import defaultdict
+from copy import deepcopy
 import functools
 import json
 import itertools
@@ -515,3 +516,32 @@ def get_cloudcare_session_data(domain_name, form, couch_user):
                 if usercase:
                     session_data[USERCASE_ID] = usercase.get_id
     return session_data
+
+
+def update_unique_ids(app_source):
+    from corehq.apps.app_manager.models import FormBase, form_id_references, jsonpath_update
+
+    app_source = deepcopy(app_source)
+
+    def change_unique_id(form):
+        unique_id = form['unique_id']
+        new_unique_id = FormBase.generate_id()
+        form['unique_id'] = new_unique_id
+        if ("%s.xml" % unique_id) in app_source['_attachments']:
+            app_source['_attachments']["%s.xml" % new_unique_id] = app_source['_attachments'].pop("%s.xml" % unique_id)
+        return new_unique_id
+
+    change_unique_id(app_source['user_registration'])
+    id_changes = {}
+    for m, module in enumerate(app_source['modules']):
+        for f, form in enumerate(module['forms']):
+            old_id = form['unique_id']
+            new_id = change_unique_id(app_source['modules'][m]['forms'][f])
+            id_changes[old_id] = new_id
+
+    for reference_path in form_id_references:
+        for reference in reference_path.find(app_source):
+            if reference.value in id_changes:
+                jsonpath_update(reference, id_changes[reference.value])
+
+    return app_source
